Ensure configured serv_list is available when fetching slots
Previously, if a cached serv_list existed then these would be the only
servers consulted. However, it is possible that the config serv_list has
been updated which means any new addresses will not be considered.

In most scenarios this is not a problem - the cluster knows the current
active topology. But in the event that all nodes are suddenly lost (for
example, the entire cluster is brought up on new addresses), the cache
presumes itself to still be the source of truth.

This commit appends our configured serv_list to the cached list (if
present), and so if all cached servers fail, it will try from the
configured list.