Fix might wrong parse the hash tag of the key
Currently, the parse_key function parses the hash tag without checking
if the right position is greater than the left position, so it might
retrieve the wrong hash tag.

For example, the whole key is expected if the key is 'A}B{C' but it will
get an empty string, and it might be redirected to the wrong slot
number. Also, it's unnecessary to find the right tag index if the left
index is not found.